requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.1.0
schedule==1.2.2
scrapy==2.13.2
